For the rare case, count the queue entries for a bucket.
If a bucket receives inserts in rapid succession,
the queue will contain a number of entries that must be processed before the thread indicates that it is done editing. The number of entries cannot be bigger than the number of second phase processing threads. Should be small number, but allow for 100+ count. Queues are by default limited to 100 entries.